chore: rename p2p_proto_v0 to p2p_proto and flatten its module hierarchy
