+++ Cargo.toml
@@ -11,7 +11,7 @@ members = [
     "crates/p2p",
     "crates/p2p_bootstrap",
     "crates/p2p_proto_derive",
-    "crates/p2p_proto_v1",
+    "crates/p2p_proto",
     "crates/pathfinder",
     "crates/retry",
     "crates/rpc",
