+++ crates/p2p/Cargo.toml
@@ -27,7 +27,7 @@ libp2p = { version = "0.51.3", default-features = false, features = [
     "tokio",
     "yamux",
 ] }
-p2p_proto_v1 = { path = "../p2p_proto_v1" }
+p2p_proto = { path = "../p2p_proto" }
 pathfinder-common = { path = "../common" }
 prost = "0.12.1"
 rand = { workspace = true }
