+++ crates/p2p/src/behaviour.rs
@@ -13,12 +13,12 @@ use libp2p::relay::client as relay_client;
 use libp2p::request_response::{self, ProtocolSupport};
 use libp2p::swarm::{keep_alive, NetworkBehaviour};
 use libp2p::{identity, kad};
-use p2p_proto_v1::block::{
+use p2p_proto::block::{
     BlockBodiesRequest, BlockBodiesResponseList, BlockHeadersRequest, BlockHeadersResponse,
 };
-use p2p_proto_v1::event::{EventsRequest, EventsResponseList};
-use p2p_proto_v1::receipt::{ReceiptsRequest, ReceiptsResponseList};
-use p2p_proto_v1::transaction::{TransactionsRequest, TransactionsResponseList};
+use p2p_proto::event::{EventsRequest, EventsResponseList};
+use p2p_proto::receipt::{ReceiptsRequest, ReceiptsResponseList};
+use p2p_proto::transaction::{TransactionsRequest, TransactionsResponseList};
 
 #[derive(NetworkBehaviour)]
 #[behaviour(out_event = "Event", event_process = false)]
