+++ crates/p2p/src/client/peer_aware.rs
@@ -4,13 +4,13 @@ use std::collections::HashSet;
 
 use anyhow::Context;
 use libp2p::{gossipsub::IdentTopic, request_response::ResponseChannel, Multiaddr, PeerId};
-use p2p_proto_v1::block::{
+use p2p_proto::block::{
     BlockBodiesRequest, BlockBodiesResponseList, BlockHeadersRequest, BlockHeadersResponse,
     NewBlock,
 };
-use p2p_proto_v1::event::{EventsRequest, EventsResponseList};
-use p2p_proto_v1::receipt::{ReceiptsRequest, ReceiptsResponseList};
-use p2p_proto_v1::transaction::{TransactionsRequest, TransactionsResponseList};
+use p2p_proto::event::{EventsRequest, EventsResponseList};
+use p2p_proto::receipt::{ReceiptsRequest, ReceiptsResponseList};
+use p2p_proto::transaction::{TransactionsRequest, TransactionsResponseList};
 use tokio::sync::{mpsc, oneshot};
 
 #[cfg(test)]
