+++ crates/p2p_proto/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "p2p_proto_v1"
+name = "p2p_proto"
 version = "0.1.0"
 edition = "2021"
 license = "MIT"
