+++ crates/p2p_proto/build.rs
