+++ crates/p2p_proto/proto/snapshot.proto
