+++ crates/p2p_proto/src/common.rs
