+++ crates/p2p_proto/src/event.rs
