+++ crates/p2p_proto/src/state.rs
