+++ crates/pathfinder/Cargo.toml
@@ -12,13 +12,13 @@ path = "src/lib.rs"
 
 [features]
 tokio-console = ["console-subscriber", "tokio/tracing"]
-rpc-full-serde = []
 p2p = [
     "dep:base64",
     "dep:p2p",
-    "dep:p2p_proto_v1",
+    "dep:p2p_proto",
     "dep:zeroize",
 ]
+rpc-full-serde = []
 
 [dependencies]
 anyhow = { workspace = true }
@@ -34,7 +34,7 @@ lru = "0.11.1"
 metrics = { workspace = true }
 metrics-exporter-prometheus = "0.11.0"
 p2p = { path = "../p2p", optional = true }
-p2p_proto_v1 = { path = "../p2p_proto_v1", optional = true }
+p2p_proto = { path = "../p2p_proto", optional = true }
 pathfinder-common = { path = "../common" }
 pathfinder-compiler = { path = "../compiler" }
 pathfinder-ethereum = { path = "../ethereum" }
